WoW renderer: split WMO code into r_wowmap_wmo.c
Extract all WMO loading/management from r_wowmap_objects.c into the new
r_wowmap_wmo.c (WOWWMOBUILD/LOAD internals, Wow_LoadWmoGroup/Model,
Wow_GetWmoModel, Wow_AddWmoInstance, Wow_WmoDoodadLocalMatrix,
Wow_QueueWmoDoodads, and coordinate/string helpers used only by WMOs).
r_wowmap_objects.c now holds doodad loading and debug-marker code only.
